0.87
버그 고치고, 노티 추가만 함

+++ Cabbage_Market/src/main/java/com/cabbage/biz/post/post/PostVO.java
@@ -24,7 +24,7 @@ public class PostVO {
 	private String sellerId;
 	private Long postCatId;
 	private String postTitle;
-	private int postPrice;
+	private long postPrice;
 	private String postContent;
 	private String postStatus;
 	private Date postRegdate;
